requirements.txt
altgraph==0.17.5
packaging==26.0
pefile==2024.8.26
pygame-ce==2.5.7
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
pywin32-ctypes==0.2.3
setuptools==82.0.1
